services/feapp/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==2.0.2
gunicorn==20.1.0
flask_restx==1.0.6
Werkzeug==2.2.2
azure-data-tables==12.4.2
python-dotenv==0.21.1
